requirements.txt
# List required packages in this file, one per line.
numpy
pyserial
pyusb
matplotlib
circular_buffer_numpy
ubcs-auxiliary
pyepics
scipy
psutil
